feat(less): Automate importing src/app and src/common less files into main.less
Any less files found in 'src/app' and 'src/common' will be automatically
imported into the base of the 'src/less/main.less' file at build time.
To achieve this the build process must make a temporary copy of the
main.less file, append a string of imports to the base of this file, use
it as the target of the recess compilation process, and then then delete
the temporary file. This may not be the most ideal solution, but it
allows users to drag and drop folders into their projects 'app' and
'common' directories without worrying about adding their less imports to
the main.less file. The advantage to including the less files as imports
is that the 'src/app' less files can still exploit any less variables
that have been defined in app.less. However, this approach may encourage
bad practices with the 'src/common' less files.