Removed adding vendor css files to head
During the build process, references to the vendor css files were placed in the head section.
However, this was an unnecessary step => vendor css files are added to the concatenated css file (row 375), resulting in errors (css files not found) in the browser. This commit removes the adding of these files to the head section.